datamodel done for MenuItem
