+++ .gitignore
@@ -1,10 +1,6 @@
-<<<<<<< HEAD
-node_modules
-=======
 # package directories
 node_modules
 jspm_packages
 
 # Serverless directories
 .serverless
->>>>>>> serverless-setup
