--- boilerplate/packages/base/menuItem/ui/index.ts
@@ -1,17 +1,0 @@
-// import './style.scss';
-// import { ReactNode } from 'react';
-// import { StyleTypes } from '@zocom/types';
-
-interface MenuItem {
-    title: string,
-    desc: string,
-    price: number,
-    category: string
-    ingredients: [],
-} 
-
-export const MenuItem = () => {
-    return (
-        <article><article/>
-    )
-}
